+++ lib/rabbitmqctl.ex
@@ -124,8 +124,8 @@ defmodule RabbitMQCtl do
   end
 
   defp process_output(output, command, options) do
-    formatter = get_formatter(command, options)
-    printer = get_printer(options)
+    formatter = Config.get_formatter(command, options)
+    printer   = Config.get_printer(options)
 
     output
     |> Output.format_output(formatter, options)
@@ -191,39 +191,6 @@ defmodule RabbitMQCtl do
   end
   defp normalise_timeout(opts) do
     opts
-  end
-
-  def get_formatter(command, %{formatter: formatter}) do
-    module_name = Module.safe_concat("RabbitMQ.CLI.Formatters", Macro.camelize(formatter))
-    case Code.ensure_loaded(module_name) do
-      {:module, _}      -> module_name;
-      {:error, :nofile} -> default_formatter(command)
-    end
-  end
-  def get_formatter(command, _) do
-    default_formatter(command)
-  end
-
-  defp get_printer(%{printer: printer}) do
-    module_name = String.to_atom("RabbitMQ.CLI.Printers." <> Macro.camelize(printer))
-    case Code.ensure_loaded(module_name) do
-      {:module, _}      -> module_name;
-      {:error, :nofile} -> default_printer()
-    end
-  end
-  defp get_printer(_) do
-    default_printer()
-  end
-
-  def default_printer() do
-    RabbitMQ.CLI.Printers.StdIO
-  end
-
-  def default_formatter(command) do
-    case function_exported?(command, :formatter, 0) do
-      true  -> command.formatter;
-      false -> RabbitMQ.CLI.Formatters.String
-    end
   end
 
   # Suppress banner if --quiet option is provided
